[Pipelines] Move get_workflow_steps to helpers
