+++ pipeline-adapters/mlrun-pipelines-kfp-common/src/mlrun_pipelines/common/pipeline_run_type.py
@@ -1,0 +1,4 @@
+class PipelineRunType:
+    run = "run"
+    build = "build"
+    deploy = "deploy"
