+++ requirements.txt
@@ -40,7 +40,7 @@ deprecated~=1.2
 jinja2~=3.1, >=3.1.3
 orjson>=3.9.15, <4
 # mlrun pipeline adapters
-mlrun-pipelines-kfp-common~=0.3.3
+mlrun-pipelines-kfp-common~=0.3.4
 mlrun-pipelines-kfp-v1-8~=0.3.3; python_version < "3.11"
 docstring_parser~=0.16
 # TODO: Remove it after moving notificaions to the server side(ML-8069)
